Fixed: WhereBuilder for Postgres
