+++ src/NzbDrone.Core/Datastore/TableMapper.cs
@@ -183,7 +183,7 @@ namespace NzbDrone.Core.Datastore
                             (db, parent) =>
                             {
                                 var id = childIdSelector(parent);
-                                return db.Query<TChild>(new SqlBuilder().Where<TChild>(x => x.Id == id)).SingleOrDefault();
+                                return db.Query<TChild>(new SqlBuilder(db.DatabaseType).Where<TChild>(x => x.Id == id)).SingleOrDefault();
                             },
                             parent => childIdSelector(parent) > 0);
         }
